+++ cypress/support/pages/channelsPage.js
@@ -99,7 +99,7 @@ export function selectChannelVariantInDetailsPage(channelName, attributeName) {
     .click();
   cy.get(SELECT_CHANNELS_TO_ASSIGN.selectChannelsForm)
     .find(BUTTON_SELECTORS.submit)
-    .click();
+    .click({ force: true });
 }
 
 export function enterHomePageAndChangeChannel(channelName) {
